Comparison of narrow type with wide type in loop condition
CWE-190
CWE-197
CWE-835

Use a loop index type that is at least as wide as partStatus.size() and cast
only where narrower API parameters are required.

Best fix here:

- In src/GenericClientListCtrl.cpp, inside
  CGenericClientListCtrl::DrawSourceStatusBar, change the loop counter from
  uint16 to uint32.
- Keep comparisons and indexing (partStatus.get(i)) in the wider type.
- For comparisons against lastDownloadingPart / nextRequestedPart (both uint16)
  and calls likely expecting 16-bit part index (reqfile->GetPartSize,
  reqfile->IsComplete), use an explicit static_cast<uint16>(i) at those exact
  use sites.

This avoids overflow in the loop condition without changing behavior for valid
part ranges.

+++ src/GenericClientListCtrl.cpp
@@ -1335,24 +1335,24 @@ void CGenericClientListCtrl::DrawSourceStatusBar(
 									? source.GetLastDownloadingPart() : 0xffff;
 		uint16 nextRequestedPart = source.GetNextRequestedPart();
 
-		for ( uint16 i = 0; i < partStatus.size(); i++ ) {
+		for ( uint32 i = 0; i < partStatus.size(); i++ ) {
 			uint64 uStart = PARTSIZE * i;
-			uint64 uEnd = uStart + reqfile->GetPartSize(i) - 1;
+			uint64 uEnd = uStart + reqfile->GetPartSize(static_cast<uint16>(i)) - 1;
 
 			CMuleColour colour;
 			if (!partStatus.get(i)) {
 				// client does not have this part
 				// light grey
 				colour = bFlat ? crFlatNeither : crNeither;
-			} else if ( reqfile->IsComplete(i)) {
+			} else if ( reqfile->IsComplete(static_cast<uint16>(i))) {
 				// completed part
 				// green
 				colour = bFlat ? crFlatBoth : crBoth;
-			} else if (lastDownloadingPart == i) {
+			} else if (lastDownloadingPart == static_cast<uint16>(i)) {
 				// downloading part
 				// yellow
 				colour = crPending;
-			} else if (nextRequestedPart == i) {
+			} else if (nextRequestedPart == static_cast<uint16>(i)) {
 				// requested part
 				// light yellow
 				colour = crNextPending;
